Fixed some issues with jungle raid porting

+++ src/main/java/gg/packetloss/grindstone/city/engine/jungleraid/JungleRaidClass.java
@@ -27,8 +27,8 @@ public enum JungleRaidClass {
         this.tntAmt = tntAmt;
     }
 
-    public short getArrowAmount() {
-        return (short) arrowAmt;
+    public int getArrowAmount() {
+        return arrowAmt;
     }
 
     public boolean hasFlintAndSteel() {
